near production release (full functionality)

+++ sql/userdata.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "UserData" (
     "token" TEXT NOT NULL,
     "discord_id" TEXT NOT NULL,
-    "metabits" DECIMAL(100,0) NOT NULL DEFAULT 0,
+    "metabits" BIGINT NOT NULL DEFAULT 0,
     "dino_rank" INTEGER NOT NULL DEFAULT 0,
     "prestige_rank" INTEGER NOT NULL DEFAULT 0,
     "singularity_speedrun_time" DOUBLE PRECISION,
